Summary - Assignment 2
How many locations are there in the data set?
Within the dataset, there are 106 locations in England.

How many service settings, context types, national categories, and appointment statuses are there?
There are 5 service settings, 3 different context types, 18 national categories and 3 appointment statuses. 

What are the five locations with the highest number of records?
Out of the 106 locations, the top five locations with the highest number of records are NHS North West London, NHS North East London, NHS Kent and Medway, NHS Hampshire and Isle of Wight and NHS South East London. It is interesting to see that these locations are not in London. London’s population makes it the largest in the UK. Therefore, one would think that there would be more records for the whole of London. Instead, we overall that the records are highest in the South of England rather than the North. Potentially this could indicate the need for possible further examination.